Some minor work on database layer
